AI Cursor
AI cursor refers to the use of eye movements to control the movement of a cursor on a computer screen or other digital displays. The proposed system focuses on providing a simple and convenient interactive mode by only using user’s facial movements.  The proposed system describes the implementation of both iris and movement of cursor according to head position which can be used to control the cursor on the screen using webcam and implemented using Python.